+++ .github/config/codeql-config.yml
@@ -1,3 +1,6 @@
 query-filters:
   - exclude:
       id: java/stack-trace-exposure
+  - include:
+      kind:
+        - alert-suppression
