+++ sql/src/main/java/org/apache/druid/sql/calcite/planner/convertlet/A2.java
@@ -17,16 +17,16 @@
  * under the License.
  */
 
-package org.apache.druid.sql.calcite.rel;
+package org.apache.druid.sql.calcite.planner.convertlet;
 
-@SuppressWarnings("lgtm[java/missing-clone-method]")
-public class A6 implements Cloneable
+@SuppressWarnings("codeql[java/missing-clone-method]")
+public class A2 implements Cloneable
 {
-  public final int a;
+  int a;
 
-  public A6(int aa)
-  {
-    a = aa;
+  @SuppressWarnings("lgtm[java/missing-clone-method]")
+  static class LG {
+    int a;
 
   }
 }
