+++ sql/src/main/java/org/apache/druid/sql/calcite/planner/convertlet/A4.java
@@ -17,9 +17,10 @@
  * under the License.
  */
 
-package org.apache.druid.sql.calcite.rel;
+package org.apache.druid.sql.calcite.planner.convertlet;
 
-public class A1 implements Cloneable
+@SuppressWarnings("codeql[clone]")
+public class A4 implements Cloneable
 {
   int a;
 }
